Move Cardano.Ledger.Tx to Shelley.Spec.Ledger.Tx
The theme behind this change is that consensus (and the node, and
components above this) will regard an Alonzo Tx to be the full Tx,
including the `IsValidating` tag.

Where previously we exposed the unvarnished `Tx` and kept the
`TxInBlock` as a ledger-internal detail, we will now be doing the
opposite - exposing the full `TxInBlock`.
