+++ alonzo/test/lib/Test/Cardano/Ledger/Alonzo/AlonzoEraGen.hs
@@ -54,7 +54,6 @@ import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.Mary.Value (policies)
 import Cardano.Ledger.ShelleyMA.AuxiliaryData as Mary (pattern AuxiliaryData)
 import Cardano.Ledger.ShelleyMA.Timelocks (Timelock (..))
-import Cardano.Ledger.Tx (Tx (Tx))
 import Cardano.Ledger.Val (adaOnly, (<+>), (<×>))
 import Cardano.Slotting.Slot (SlotNo (..))
 import Control.Iterate.SetAlgebra (eval, (◁))
@@ -74,6 +73,7 @@ import Plutus.V1.Ledger.Api (defaultCostModelParams)
 import qualified PlutusTx as P (Data (..))
 import qualified PlutusTx as Plutus
 import Shelley.Spec.Ledger.PParams (Update)
+import Shelley.Spec.Ledger.Tx (Tx (Tx))
 import Shelley.Spec.Ledger.TxBody (DCert, TxIn, Wdrl)
 import Shelley.Spec.Ledger.UTxO (UTxO (..))
 import Test.Cardano.Ledger.AllegraEraGen (genValidityInterval)
