+++ cardano-ledger-core/cardano-ledger-core.cabal
@@ -55,7 +55,6 @@ library
     Cardano.Ledger.SafeHash
     Cardano.Ledger.Serialization
     Cardano.Ledger.Slot
-    Cardano.Ledger.Tx
     Cardano.Ledger.Val
 
   build-depends:
