+++ shelley/chain-and-ledger/shelley-spec-ledger-test/src/Test/Shelley/Spec/Ledger/Shrinkers.hs
@@ -8,7 +8,6 @@ module Test.Shelley.Spec.Ledger.Shrinkers where
 import Cardano.Ledger.Coin
 import qualified Cardano.Ledger.Core as Core
 import Cardano.Ledger.Slot
-import Cardano.Ledger.Tx
 import Cardano.Ledger.Val ((<+>), (<->))
 import qualified Cardano.Ledger.Val as Val
 import Data.Foldable (toList)
@@ -25,6 +24,7 @@ import Debug.Trace (traceShowId)
 import Shelley.Spec.Ledger.BlockChain
 import Shelley.Spec.Ledger.PParams
 import Shelley.Spec.Ledger.Scripts
+import Shelley.Spec.Ledger.Tx
 import Shelley.Spec.Ledger.TxBody
 import Test.QuickCheck (shrinkIntegral, shrinkList)
 import Test.Shelley.Spec.Ledger.Utils (ShelleyTest)
